requirements.txt
pytelebot==0.0.4
